backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
pandas>=2.2.0
google-genai>=1.0.0
reportlab>=4.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
